feat[be]: Add index in elasticsearch  and modify create_project_index function
